Move Nailed::Config to separate file
